fix: move type statement of handlers to child component to use them.
